Add TomlContentPolymorphicSerializer, add TomlEncoder.toml/TomlDecoder.toml, add TomlNativeWriter, add TomlConfig.explicitNulls, add contracts to various type checks, add initial support for sequential parsing.
