+++ benchmark/src/jmh/kotlin/test/BytecodeChecker.kt
@@ -3,11 +3,6 @@ package test
 import kotlinx.serialization.Serializable
 
 @Serializable
-sealed class B1 {
-    abstract val i: Int
-}
-
-@Serializable
-data class M1(
-    override val i: Int
-) : B1()
+data class M(
+    val s: String?
+)
